chore: organize graph pipelines
