+++ TASKS.md
@@ -2,4 +2,5 @@
 
 - [x] Document usage examples for `crawlGraph` and `exploreGraph` in `@civ7/plugin-graph`.
 - [x] Add CLI integration tests verifying the `crawl` and `explore` commands invoke plugin pipelines.
+- [x] Clean up file/directory structure for CLI and plugin-graph packages.
 - [ ] Improve error handling and logging in graph plugin pipelines.
