+++ packages/plugins/plugin-graph/TESTING.md
@@ -1,0 +1,17 @@
+# Testing
+
+This package uses [Vitest](https://vitest.dev/).
+
+## Running tests
+
+```bash
+pnpm -F @civ7/plugin-graph test
+```
+
+## Current smoke tests
+
+- `crawlGraph` and `exploreGraph` pipelines build graphs and render SVG.
+
+## Suggested future tests
+
+- Edge cases for graph export and renderer options.
